WIP - Refactor and test of clickRateIt function
